Initial commit I started by taking my code from my Poisson_PINN project.
I rewrote the network class for the new setup. I also rewrote the main
file, the setup file reader, and setup.txt.

I still need to write a data loader.

I also need to update the plotter.

I also want to add a setting that controls how many derivatives the
N network is a function of (and make corresponding changes to the
Coloc loss function). This should make the code a lot easier to use when
switching between diff eqs.