Implemented data loader
I also added some new settings to control the number of testing
and training points. Currently, the collocation and data
coordinates are the same. I may want to look into using different
points for these.

Also, this code really needs batching. I plan to add some form of
batching in the next commit or two. There's simply too much data to
not do batching. This will require using the same number of
collocation and data points. I'm okay with that, however (it's what
they did in the paper after all).

In my next commit, I plan to fix the plotter.

After that, I want to make the number of derivatives of u that
N accepts a setting. This change should make my code wayyyyyyy more
flexible. From there, the bulk of the code should be done and I should
be able to move into testing (and finding the dozens of bugs that
I'm sure I've introduced).