Rewrote plotter, sorta
I partially wrote the plotter. I realized, however, that I can't
implement this function as is.

The reason is that I need the coordinates of every data point to
make this thing work. I need to modify the way that the data
loader works, so that it returns all the data points + all data values
in addition what it already returns. I can then feed that junk into
the plotting functions. Woo-hoo, 8 return arguments.

I also wrote a new function whose purpose is to calculate the PDE
residual at a set of points. I realized that this code appeared
twice in my project (once in residual part of the plotter, and
again in the Coloc loss function in Network). This new function
eliminates code repetition.

My first goal for the next commit is to implement the plotter using the
changes discussed above. After that, I want to make the number of
derivatives of u used as inputs to N into a setting.