Got plotter running + made lots of tweaks
I made the plotter work! To do this, however, I had to rewrite a lot of
the data loader. In particular, decided that 6 return arguments was
dumb, especially since I was going to add more. Thus, I switched to
returning a data container object. This object has a member for
each return argument. This reduces bugs because I don't have to keep
track of what goes in what slot.

With this, and a lot of rewrites in plotter, I was able to get the
thing to plot. Woo-Hoo.

I also added a setting to decide if we should plot the results or not.
I should add that it takes about 3 minutes for the plot to work.

In my next commit, I want to clean stuff up and make the number of
derivatives that N depends on a setting.