Updated plotter + added folders
I added new folders for Data and Saves. This should help organize the
project a little better. I made corresponding changes to the main
function so that it loads/saves in the right location.

I also rewrote the plotter. I decided that I should have an "Error"
plot in addition to the others. Thus, I added a 4th subplot and
changed the ploting routine to plot a 2x2 grid of subplots. Making
this work such that it didn't look like absolute crap took a lot of
time. What a surprise, it's matplotlib. Anyway, it works (I think).

I my next commit, I plan to reorganize the setup file so that the
comments go near the corresponding settings.

After that, I plan to make a sequence of major changes. I want to make
the code able to run in "PINN" mode. In this mode, the NN network is
fixed. The BC, IC, and Collocation losses are used to find a solution
that satisfies the learned PDe (given by the NN network). This will
replace the existing "enable NN learning" setting. This change is
going to take some time to implement, so expect a sequence of commits.