Added IC, BC to data loader
The data loader now loads IC coordinates + Data, as well as coordinates
for the upper and lower bounds of the spatial domain (at each possible
time), which can be used to enforce periodic BCs.

This is the first step twoards allowing my code to run the PINN
algorithm once the NN network is trained. Expect new loss functions in
the next commit.