Added Loss_Function, Test_Train files
I added new files for the loss functions and testing/training functions.
I did this just to make the code more organized/easier to maintain
(the files were getting a tad long).

I still have not tested anything. The code will almost certainly break
if run. DO NO RUN THIS CODE. In the next commit, I plan to make the
data loader work with the Mode argument (basically set IC/BC data to
NONE if in Discovery mode, and the Testing and Training point stuff to
NONE if in PINNs mode). I will then test the crap out of the code to see
if the new setup works. Assuming I can get the code running correctly in
a reasonable amount of time, I'll try and update the periodic BCs
function so that it can evaluate a variable number of derivatives.
Hopefully once that is done, the code will settle down again (no major
changes for some time).