Fixed bugs, updated Data_Loader
I updated the Data_loader so that it's behavior depends on which
mode the code is in. Anything that's not used in the current mode
is set to None... hopefully that will throw off errors if I try to
use some values in one mode that aren't used for that mode (built in
debugging... I guess. Or something like that).

I also debugged a lot of the code/got it to actually run. I haven't
done testing on any of the new code, so I have no idea if it's running
correctly. However, it can at least run when in "Discovery mode". I
haven't tried anything else yet. Expects lots of testing/debugging
in the next few commits.

I also rewrote a bunch of comments/reorganized some of the code.

Expect an updated Periodic_BC_Loss function in the next commit. After
that, I need to debug everything. However, once that's done, the code
should remain stable for a while, as it'll be able to do everything
that it needs to do for now.