Added BC loss description, fixed bugs
I added a proper description to the BC loss function. I also made
tweaks to other comments throughout the code.

I also redid the way that the Data_Loader works. The Data_Container is
now an empty container (just like the Setup Container). I selectively
add elements to this container depending on what mode we're in. I like
this approach because it's bug proof. If I try to use a variable that
is specific to the PINNs mode while in Discovery, then the Python
interpreter will catch it and throw an error. With the old approach,
all the variables always existed, even if they weren't used in one mode.
Thus, this change is purely to make the code more robust/help ensure
that it's doing what it does.

The code appears to run in both PINNs and Discovery mode. I need to
run a lot more tests, and probable clean up a bunch of the code and
comments, but I'm pretty happy with the state of things.