Added a timing class
I plan to rewrite the way that my code uses the NN class (passing a
batch of points at once rather than one at a time). I expect this to
increase performance (especially when running on a GPU), but want to
be able to measure that improvement (as well as any future performance
improvements). Thus, I wrote a simple timing class. I put it in a new
file and then added bits to the main function to use the timers. It
appears to work.

Expect changes to the Loss_Functions (batches of points rather than
one at a time) in the next commit.