Batched computations yield enormous gains
I updated the Data_Loss, Collocation_Loss, and PDE_Residual functions
in the Loss_Functions file. These functions now pass points through the
NN in batches (which is apparently how torch's module expects them to
work). This was a little tricky with regards to computing the
derivatives. In particular, I think torch's autograd.grad is really
designed for computing the derivatives of scalars. It works recursively
to implement the chain rule, and uses non-scalar outputs in these
recursive steps. This means that it can compute the derivative of a
non-scalar outputs variable. However, based on the complete lack of
documentation explaining what this actually does (all existing
documentation just states what size everything needs to be... not
what it's actually finding), I think this use case was really only
intended for pytorch itself. It took me a long time to figure out how to
make this function compute what I wanted it to (a 2 column tensor whose
ith row holds the derivative of the ith output with respect to the ith
x, t coordinates). I got it to work though (I think).

With these changes, the runtime improved by about 2 orders of magnitude.
Before, it took nearly 8 seconds per epoch. It now takes about .05. Wow.
The only real difference is that I eliminated a python loop (essentially
replacing it with a C one) and improved memory locality with batching.
It's amazing how much that improves runtime. It's almost like
interpreted languages are slow and shouldn't be used for HPC or
something (yes, this is a salty sentence. I miss C++).

I plan to rewrite the other loss functions (For PINNS mode) using this
batched approach as well. Expect those changes in the next commit.
After that, I may begin working on implementing an extraction mode...
or maybe I'll try to profile the code to gauge if it's running
sufficiently fast now, or if I should look for other ways to improve
runtime. The todo file explains some of my other plans for the code.