Added multi-index test code to commit
I've spent most of the day writing (and past few days thinking about)
implementing a function which essentially finds the set of "distinct"
multi-indicies of a specified order whose sub-indicies take values in
some specified range. This is important because it's how we will
populate the "library" of possible terms for extraction mode. Here,
two multi-indicies are "distinct" if one can not be permuted into the
other. This defines an equivalence relation.

Why do we define "distinct" this way? Because scalar multiplication
commutes. This code is going to be used to essentially find polynomial
terms. the term x*y*x is not distinct from x*x*y... multiplication
commutes. The definition of "distinct" above helps implement this fact.

I am committing the Multi_Index_Test file just so that I can track
changes. This file will eventually turn into the extraction mode code.
Expect this file to be renamed + substantially expanded in the next few
commits.