Redid data loading + made other tweaks
I totally rewrote the way that the code sets up collocation and data
points. To begin, the old code was hardcoded to use the same
collocation and data points. This meant that all collocation points
existed on the grid of data points. I wanted them to be random. I also
needed to add the ability to add "extraction points" for extraction
mode. This prompted me to rewrite most of the data loader function
and make big changes to the setup file.

I added a lot of mode specific settings (see setup.txt), such as
mode specific controls on the number of collocation and data points
in training and testing. All collocation points (as well as extraction
points) are now randomly generated within the problem domain.

I eventually want to change how the code selects the problem domain.
In particular, I think this should be something that gets set in the
Data_Setup function. It will also have to adapt with the number of
spatial dimensions... expect those changes in the nearish future.

These changes forced me to rewrite a bunch of the main function
as well, and to change some things in the Setup File Reader (to account
for the new settings). This is a big commit, but I think it's working.

In the next commit, expect to see some tweaks to the way that the
data container is set up (namely the problem domain bounds/number
of variables in the solution).

After that, I'll actually implement extraction mode (this time with
everything set up correctly).