Wrote Generate_Library + assertions + fixed bugs
I wrote the Generate_Library function. I believe it is working (though
I have not yet tested this. I should probably write a unit test to do
just that. It would work with a super simple 1 layer NN which is
initialized to always output 1 (weight matrix set to 0, bias vector
set to 1).

Writing the function made me realize that the Evaluate_u_derivatives
function had a bug in it! What I was calling "num_derivatives" was
really num_derivatives + 1. I fixed this / made minor tweaks to
accommodate this change.

I also added assertions to the recursive functions in Multi_Index_Test
to make sure that the inputs are in the right range.

I still need to write the actual extraction step (with least squares
and all)... but the hard part of the code is done.

In my next commit, I think I will rename Multi_Index_Test to
Extraction. I also want to start writing unit tests for a lottttt of
the code. like a lot. I guess that's what I'll do on Friday. After that,
I plan to rewrite a bunch of the code with an arbitrary number of
spatial dimensions in mind (rather than just 1). This should allow me
to generalize the code to other problems without much effort.