Wrote more tests!
I wrote tests for the rest of the Loss_Functions (except for
Collocation_Loss, since this function is basically just a wrapper for
PDE_Residual.... Expect a test for that function in the next commit).

In my next commit, I plan to write tests for the functions in
PDE_Residual.py.

On Wednesday, I'll write the rests of the tests (mainly for stuff in
Extraction). That should wrap up my unit testing plans (I don't think
it makes sense to test the rest of my code?... since most of it is
either very simple (test_train, timing), so specific to the problem
being studied that it's not worth testing (plotter, data setup), or is
not "mission critical" to the code giving correct values (timing,
plotter)).