Added optimizer setting
I added a setting to control which optimizer we use. It only accepts
two options right now: Adam and LBFGS. Switching to LBFGS forced me to
rewrite some of the test/train code. However, the important thing is
that it works now.

With only 100 epochs of LBFGS on top of a trained (using 4000 epochs
of adam) u_NN and N_NN, I was able to drop the error by about an order
of magnitude. Amazingly, when I extracted the refined PDE, it got the
right PDE and got the coefficients to within 2-3% :D :D

Further, I was able to increase the PDE term order to 4 (I didn't test
any higher) and it still consistently worked :D :D :D :D