Added a Data type setting
I spent most of the day trying to learn about LBFGS and quasi Newton
optimization methods. However, I decided to take a break from that to
add a new feature to the code. In particular, I added a setting which
controls the data type of all tensors in my code.

I had hard coded everything to be a single precision float before.
However, I was quickly approaching machine precision (I think). I,
therefore, decided to add a setting to control precision (just to see
if it'd help). This change forced me to modify a lot of the code (which
had been hard coded to work with float32's).

Anyway, the change worked.