Added support for Rational ReLU function
I added support for the Rational function approximation to the ReLU
activation function. To do this, I had to implement a custom class.
I wrote new tests to verify that the new activation function works
as intended.

I removed support for ReLU. The idea is to use this approximation in its
place. ReLU was ill posed for this problem anyway (its second derivative
is zero.... it can't approximate second derivatives).

Everything seems to work, though the Rational activation function is
very very slow. I need to figure out way to improve performance. This
may be difficult, since every evaluation involves division, and division
is inherently slow.

I plan to profile the new Rational_ReLU function. I want to see how
long it takes to do forward evaluation in a rational vs tanh network.
If the rational network is way slower, then I should be able to improve
performance by tweaking how the rational class's forward method works.
If not, then the bottleneck must be in derivative computations. In this
case, I guess I'll have to figure out how to make the derivative
calculation faster (or rather, how to get torch to be smarter about
how it evaluates the derivative). I also want to know how well the
rational network performs if I run a lot of epochs. There's a lot of
experimenting to be done.