Improved rational runtime + added profiling
I created a new file, called Profiling.py, in the test directory. This
is, as the name suggests, used to profile various bits of code.
Currently, it only profiles the rational activation function. I may
expand it later.

Through profiling, I found some ways to make the rational activation
function evaluate slightly faster (namely by using Horning's rule).
It's still far slower than sigmoid or tanh, but I guess progress is
progress. I tried a lot of things, few of them worked.

I also tried running the code with N accepting 3 derivatives of u.
Amazingly, after 2000 iterations of Adam, and 500 of LBFGS, the
code was able to extract the correct PDE with coefficients only
~1% off. This worked even if I kept the term degree at 4 (5 sometimes
worked, but not always). This is just as good as I got when I set
N to accept 2 derivatives. So far, so good.