Improved LASSO (it's still bad)
I modified my code's LASSO algorithm. I changed the function "Lasso" to
"Lasso_Selection". Now, the algorithm first uses LASSO to essentially
figure out which terms it should pay attention to. In particular, after
running LASSO, we make note of which coefficients are non-zero. We then
perform least squares on the corresponding columns of the Library. In
essence, we use LASSO to identify the key terms in the dynamics, and
Least squares to identify the corresponding coefficients. This approach
sorta works, though not terribly well. With that said, it can usually
extract the correct equation and the coefficients are very good (as
good as with thresholded). Thus, this marks a major improvement.