Added noise option
I added an option to add in noise, which is added in by the Data_Setup
function. How does this work? First, I calculate the standard derivation
of the input data. I then multiply this by the Noise Proportion. That
number becomes the standard deviation of a normal distribution with mean
0, which then calculates noise/adds it onto the input data.

I kinda want to add an option to "save data", since the current setup
recalculates the noise each time that the data is loaded (this may not
be a problem, to be honest... I need to think about it more).