Added a "save plot" setting
I added a setting to control if the plot should be saved. The code
now generates a plot if the user wants to display the plot, or if
they want to save the plot. the plot is saved to a new directory called
"Figures", and uses the same file name as the save file name.