Updated Loss printout
I totally redid how the code reports losses. First of all, I want the
code to print the test and train losses... this should help me determine
if the networks are actually learning the underlying PDE, or just
memorizing the data.

I experimented with a few ways to implement this, but ran into trouble
with all that tried to print on one line. In the end, I decided to print
 on two lines. I also realized that evaluating the loss functions at
all test/train points is reasonably expensive. Thus, I decided to change
how often it reports losses (rather than doing so every epoch, it now
only does so periodically). This change should (slightly) improve
runtime.

I only implemented the changes for Discovery mode. I may add them in for
PINNs in the future.