Loaded optimizers now use settings learning rate
Before this commit, if I loaded an optimizer from save, it would use
the learning rate that was set in the save. This was annoying, since
it basically meant that I couldn't change the learning rate between
runs (on the same data) without changing the optimizer. I set it up this
way simply because I didn't know how to change the learning rate in
the optimizier object.

I spent a few hours learning about how pytorch implements its optimizer
objects, which taught me how to modify the learning rate (and what
parameter groups are).

Thus, the optimizer should ALWAYS use the learning rate specified in
Settings.txt, regardless of if the optimizer is loaded or new.