Added support for dropout + fixed some bugs
I fixed a few bug in main, and added support fo dropout. I haven't
really tested it. I want to do that on a gcp machine.

I also plan to rename the u network as the sol network and the N network
as the PDE network. These names are less ambiguous. Expect that in the
next commit (maybe).