Added periodic redo of collocation points
I added a setting which periodically regenerates the collocation points
(both for testing and training). The idea here is to change up what data
the network is training on, so that it doesn't just memorize one set of
data.

So far, the results have been promising. I plan to try it on a GCP
machine.