Added a file to plot rational network
This is used to visualize the rational NN's when trained. I don't plan
to use this code often... but it was a pain to write, so I want a log
of it in case I need to do something similar in the future.

I also fixed a bug in Settings_Reader (didn't close the file.. oops).