Added Recursive Feature Extraction algorithm
I added a new extraction algorithm called "Recursive Feature
Extraction". This finds a sequence of progressively sparser least
squares solutions to Ax = b. To begin, we first solve Ax = b in a least
squares sense. We then use a criterion (which you need to select) to
pick the "least important" feature (component of x). We remove that
column of A and then repeat. This process continues until we have just
one feature remaining (in theory, the "most important" feature). At each
step, we record the least squares solution and its residual.

In my case, I define the "least important" feature as the one with
the smallest magnitude. In general, this is not a principled approach.
However, it can be if we precondition the matrix A. To see this, let
R(x) = ||Ax - b||_2^2.
The least squares solution x finds the global minimum of R(x). A little
algebra shows that R(x) is a quadratic function of each component of x.
In particular, if A is an n by m matrix, then
R(x) = sum_{i = 1}^{n} ( (sum_{j = 1}^{m} A_{i,j} x_j) + b_j )^2
We can approximate (in this case, exactly), the value of R(x) if we were
to remove the kth component of x, which corresponds to changing x to
x - x_k e_k. In particular,
R(x - x_k e_k) = R(x) + (dR(x)/dx_k)x_k + (1/2)(d^2R(x)/dx_k^2) x_k^2
Since the least squares solution is at a minima, and since R is smooth,
dR(x)/dx_k = 0. Therefore, we must have
R(x - x_k e_k) - R(x) = (1/2)(d^2R(x)/dx_k^2) x_k^2
A little algebra shows that
(d^2R(x)/dx_k^2) = sum_{i = 1}^n A_{i,k}^2 = ||A_k||_2^2
In other words, (d^2R(x)/dx_k^2) is the square of the L2 norm of the kth
column of A. If we precondition A such that each column has unit L2 norm
(by dividing the kth column by ||A_k||^2), then we get
R(x - x_k e_k) - R(x) = (1/2)x_k^2
We define the least important feature as the one whose removal will
increase the residual by the least amount. The above analysis indicates
that if A is preconditioned, then this should roughly correspond to the
feature whose magnitude is smallest [^1]. This is what my algorithm does.
In a future commit, I plan to rank the potential solutions by how much
the residual increased (in a residual sense) by removing the least
important feature from that solution. Expect that in the next commit.

[^1]: this is only approximate, since
we really want to know the following: if we were to remove one feature
and solve the resulting least squares problem - a constrained version
of the original one - then which feature would increase the residual by
the least. This is not the same as what we've done. We've looked at
what would happen to the residual if we removed the kth component from
the least squares solution, x, keeping all other components fixed.
Solving the least squares problem with the constrained x_k = 0 would, in
general, give different values for all other features, meaning that the
change in residual would not match the prediction given by the 2nd order
Taylor approximation. With that said, the Taylor approximation gives a
reasonable and computationally cheap estimate (as the alternative would
involve solving a least squares problem for each remaining feature. Given
n features, this would involve O(n^2) least squares solves, which is
computationally unacceptable).