Added ranking algorithm.
I added an algorithm to rank the most likely PDEs. How does this work?
First, for each PDE, we find the relative increase in Residual from
removing the least important component of that PDE (how much the
residual changed, in a relative sense, in going from this PDE to the
next sparsest one).

I print out the top 5 most likely PDEs.