Updated timing output
Previously, my code only displayed the runtime in seconds. I often found myself converting this into minutes to get a sense of how long it took. This prompted me to just change the output. It now displayes in minutes, seconds (as well as the total runtime in seconds). Hopefully this will be more useful.