Extraction mode is now GPU enabled, fixed time bug
I finally got access to a GPU machine. This allowed me to test my
code on a GPU. Much to my surprise, Discovery mode actually worked
.... the first time.... without crashing. I've literally never
tested this code on a GPU before (or, really, any code)... so that
was a big surprise. Unfortuneatly, Extraction mode didn't work.

However, I was able to fix the problem. I had to set the device
attribute on the diu_dxi and du_dt tensors to the Device argument
in the Extraction function. oops. I also found out that you need
to convert a tensor to the CPU before converting it to a numpy
array. Big oops. Anyway, I was able to fix those issues. Woo-hoo.

I also fixed a small formatting bug in printout for the runtime
in Discovery mode.