Removed datatype setting, Sigmoid activation
I never used the data type setting. As such, I decided to remove it (in the interest of simplifying the settings file). This forced me to rewrite a bunch of code (or really, tweak a bunch).

I also brought the settings reader into 2022. I adapated my settings readter from PDE_LEARN. This allowed me to clean up the code in settings reader... a lot.

Finally, I decided to change the options for networks. I never used the Sigmoid activation function. I decided to replace it with sine (allowing for "SIRENS"). I also added support for the SGD optimizer. Idk if I will use it, but why not. I guess?

In my next commit, I plan to make a few tweaks to the plotter. Expect new datasets soon.