Updated plotter + some matlab scripts
I made some big changes to the plotter (specifically the dataset plotter). It was
seriously bugged before (or, rather, matplotlib is bugged). Basically, the way
that matplotlib determines the range of a dataset it's printing is terrible... if you
let it choose, it will sometimes just delete values form the dataset/not print them.
I noticed this when trying to plot the Burgers_Sine dataset. I fixed it by manually
specifying the levels as a numpy array.

I also updated some bits of the matlab scripts (namely the heat ones) to
adhere to my naming scheme / save their files to the Data directory.