Removed 2d data sets, added KG data set
I do not plan to make PDE_READ work in 2d in the immediate future. As
such, I decided to remove the 2D data sets. I did, however, take
advantage of the fact that PDE_READ can now learn equations of the form
          D_t^m U = N(U, D_x U, ... , D_x^n U)
In particular, I added the Klein-Gordon equation (which looks like the
wave equation plus an exponential term).