Added Batch Normalization option
I adeed an option to add a batch normalization layer before the N network. I realized that for some of my data sets, the solution lies in the range [0, 1], while it's 4th derivative lies in the range [-160,000, 160,000]. That's a big range. I figured that a batch normalization layer can help fix such discrepencies. Batch Normalization appears to harm PDE-READ in some cases (such as the beam equation. With BN, it can't work with 50% noise, while it can without), and help it in others. As such, I added it as an option for future experimentation. At the very least, I think it can be a handy tool for when PDE-READ fails to converge to the data set. Unfortuneatly, since this change added a setting, I had to update Settings.txt and /Plot/Settings.txt.

I also slightly adjusted the PDE printout to be a little easier to read.