Start of Static Beam branch
This branch exists to run the ONR static beam tests. However, I'm also using it to test some bigger changes that I want to make to PDE-READ (and other projects) in general. Currently, PDE-READ genetates the data sets that it tests on. In particular, the Data_Setup.py file reads in a .mat file (produced by one of my MATLAB scripts), turns it into a set of numpy arrays, adds a desired level of sparsity and noise, and then returns the resulting sparse/nosity data set. There is a big problem here: this code only works if the data sets are produced by my matlab scripts. If the data is in another format, everything breaks down.

I want to change this by replacing the "data set generation" part of PDE-READ. To do this, I intend to write a new "Read_DataSet.py" file to replace the existing "Data_Setup.py" one. This file will read a file from the Data/DataSets directory. Each file in this directory will contain two certialized arrays, one to store the input coordinates, and another to store the target values. PDE-READ does not care what process made the files in DataSets file, only that they use this common format. The code that actually generates the files in Data/DataSets can be problem-specific/custom. This allows PDE-READ to read in arbitrary data without modifying its code. The code that actually makes the files in Data/DataSets will be located in the Data directory.

This process is fairly similar to how I spun off the Plot portion of PDE-READ; like data-set generation, plotting is a problem-specific problem that does not belong within PDE-READ. In essence, this change makes PDE-REAd more focused on its core task - learning a PDE from data.

This commit does not contain everything I just described above. Instead, this commit only changed the directory structure. The next few commits will implement all of the changes I list above.

In addition to this change, I reworked the Matlab directory (which is now called "MATLAB"). The new MATLAB directory contains two sub-directories, Scripts and Data. The scripts file houses the .m files that used to be in the Matlab directory. The Data sub-directory is literally the old Data directory.